moved the SocialNetworks component and style files
