+++ src/components/Navbar/components/MobileMenu/MobileMenu.tsx
@@ -1,7 +1,7 @@
-import { IconMenu } from './components/IconMenu/IconMenu.tsx';
-import '../../../../styles/Navbar/MobileMenu.css';
-import { SideMenu } from './components/SideMenu/SideMenu.tsx';
 import { useState } from 'react';
+import { IconMenu } from './components/IconMenu/IconMenu.tsx';
+import { SideMenu } from './components/SideMenu/SideMenu.tsx';
+import '../../../../styles/Navbar/MobileMenu/MobileMenu.css';
 
 export const MobileMenu = () => {
     const [menuIsOpen, setMenuIsOpen] = useState(false);
