+++ src/components/Navbar/components/SocialNetworks/SocialNetworks.tsx
@@ -1,7 +1,7 @@
 import { XIcon } from './components/XIcon/XIcon.tsx';
 import { LinkedinIcon } from './components/LinkedinIcon/LinkedinIcon.tsx';
 import { GithubIcon } from './components/GithubIcon/GithubIcon.tsx';
-import './SocialNetworks.css';
+import '../../../../styles/Navbar/SocialNetworks/SocialNetworks.css';
 export const SocialNetworks = () => {
     return (
         <section className="socialNetworks">
