--- src/components/Navbar/components/SocialNetworks/components/LinkedinIcon/LinkedinIcon.css
@@ -1,6 +1,0 @@
-/*.linkedinIcon{*/
-/*    border-radius: 8rem;*/
-/*    border: transparent;*/
-/*    width: 3.25rem;*/
-/*    height: 3.25rem;*/
-/*}*/
