--- src/components/Navbar/components/SocialNetworks/components/XIcon/XIcon.css
@@ -1,6 +1,0 @@
-/*.xIcon {*/
-/*    border-radius: 8rem;*/
-/*    border: 2px solid white;*/
-/*    width: 3rem;*/
-/*    height: 3rem;*/
-/*}*/
