--- src/styles/Navbar/SocialNetworks.css
@@ -1,26 +1,0 @@
-.socialNetworks {
-    display: flex;
-    justify-content: center;
-    margin-top: 20px;
-    gap: 0 6rem;
-    height: fit-content;
-
-    a {
-        height: fit-content;
-    }
-
-    .xIcon, .githubIcon {
-        border-radius: 8rem;
-        border: 2px solid white;
-        width: 3rem;
-        height: 3rem;
-    }
-
-    .linkedinIcon {
-        border-radius: 8rem;
-        border: transparent;
-        width: 3.25rem;
-        height: 3.25rem;
-    }
-}
-
